% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rodger.R
\name{f_e_alpha}
\alias{f_e_alpha}
\title{Rodger F critical values}
\usage{
f_e_alpha(v1, v2, alpha)
}
\arguments{
\item{v1}{numeric, numerator degrees of freedom}

\item{v2}{numeric, denominator degrees of freedom}

\item{alpha}{numeric, desired E_alpha}
}
\value{
numeric, Rodger's critical F for mutually orthogonal linear contrasts
}
\description{
Rodger F critical values
}
\examples{
f_e_alpha(2,12,.05)
}
